Moved all vcd native cluster utilities from cluster.py to vcdbroker.py
